GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - feature/v0.2 (refs/gitbutler/feature/v0.2)
   branch head: e96e62e3a17a9a4ecc023b2a27ce5fa2d07d32ea
   - library/src/main/java/com/ding1ding/jsbridge/WebViewJavascriptBridge.kt
   - app/src/main/java/com/ding1ding/jsbridge/app/MainActivity.kt
   - library/src/main/java/com/ding1ding/jsbridge/Logger.kt
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch